Configure as client application module (with `use CharonOauth2` macro)
- configure as module via __using__ macro instead of application environment
- generate models via macro's
- store client.secret encrypted with key derived from Charon base secret
- store grant.code as hmac with key derived from Charon base secret
- restructure test support code
- impr vs code tasks
