+++ test/support/my_app.ex
@@ -1,0 +1,2 @@
+defmodule MyApp do
+end
