+++ test/support/my_app/charon_oauth2.ex
@@ -1,0 +1,3 @@
+defmodule MyApp.CharonOauth2 do
+  use CharonOauth2, MyApp.CharonOauth2.Config.get()
+end
