+++ test/support/my_app/repo.ex
@@ -1,4 +1,4 @@
-defmodule CharonOauth2.Repo do
+defmodule MyApp.Repo do
   @moduledoc false
   use Ecto.Repo,
     otp_app: :charon_oauth2,
